requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=vis --resolver=backtracking
#
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
fonttools==4.40.0
    # via matplotlib
graphviz==0.8.4
    # via mxnet
idna==3.4
    # via requests
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.7.1
    # via
    #   mnist_style (pyproject.toml)
    #   seaborn
mxnet==1.9.1
    # via mnist_style (pyproject.toml)
numpy==1.22.4
    # via
    #   contourpy
    #   matplotlib
    #   mnist_style (pyproject.toml)
    #   mxnet
    #   pandas
    #   seaborn
packaging==23.1
    # via matplotlib
pandas==1.5.3
    # via
    #   mnist_style (pyproject.toml)
    #   seaborn
pillow==9.5.0
    # via
    #   matplotlib
    #   mnist_style (pyproject.toml)
pyparsing==3.1.0
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2023.3
    # via pandas
requests==2.31.0
    # via mxnet
seaborn==0.12.2
    # via mnist_style (pyproject.toml)
six==1.16.0
    # via python-dateutil
urllib3==2.0.3
    # via requests
